requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.2
contourpy==1.3.3
cryptography==46.0.7
cycler==0.12.1
fastapi==0.135.3
ffmpeg-python==0.2.0
filelock==3.20.0
fonttools==4.61.1
fsspec==2025.12.0
future==1.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.30.3
google-api-python-client==2.194.0
google-auth==2.49.2
google-auth-httplib2==0.3.1
google-generativeai==0.8.6
googleapis-common-protos==1.74.0
greenlet==3.4.0
grpcio==1.80.0
grpcio-status==1.71.2
h11==0.16.0
httplib2==0.31.2
idna==3.11
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
networkx==3.6.1
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
opencv-python==4.12.0.88
packaging==25.0
pillow==12.0.0
polars==1.36.1
polars-runtime-32==1.36.1
proto-plus==1.27.2
protobuf==5.29.6
psutil==7.1.3
psycopg2-binary==2.9.11
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.13.0
pydantic_core==2.46.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.26
PyYAML==6.0.3
requests==2.32.5
router==0.1
scipy==1.16.3
setuptools==80.9.0
six==1.17.0
SQLAlchemy==2.0.49
starlette==1.0.0
sympy==1.14.0
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.3
triton==3.5.1
typing-inspection==0.4.2
typing_extensions==4.15.0
ultralytics==8.3.237
ultralytics-thop==2.0.18
uritemplate==4.2.0
urllib3==2.6.2
uvicorn==0.44.0
